implemented google fact check tools api
implemented google fact check tools api to enhance the model to be more accurate. with more trusted data to draw from the predictions will be more reliable, but this also introduces many security concerns because the model is not closed anymore. next implementations will be a couple of security measures.